requirements-to-freeze.txt
flask
python-dotenv
flask-sqlalchemy
flask--migrate
flask_restful
gunicorn
PyMySQL
-e .
factory_boy
